move test to separate "op" file
