+++ core/rawdb/accessors_chain_test.go
@@ -32,7 +32,6 @@ import (
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/params"
 	"github.com/ethereum/go-ethereum/rlp"
-	"github.com/stretchr/testify/require"
 	"golang.org/x/crypto/sha3"
 )
 
@@ -816,36 +815,6 @@ func TestReadLogs(t *testing.T) {
 			}
 		}
 	}
-}
-
-func TestParseLegacyReceiptRLP(t *testing.T) {
-	// Create a gasUsed value greater than a uint64 can represent
-	gasUsed := big.NewInt(0)
-	gasUsed = gasUsed.SetUint64(math.MaxUint64)
-	gasUsed = gasUsed.Add(gasUsed, big.NewInt(math.MaxInt64))
-	sanityCheck := (&big.Int{}).SetUint64(gasUsed.Uint64())
-	require.NotEqual(t, gasUsed, sanityCheck)
-	receipt := types.LegacyOptimismStoredReceiptRLP{
-		CumulativeGasUsed: 1,
-		Logs: []*types.LogForStorage{
-			{Address: common.BytesToAddress([]byte{0x11})},
-			{Address: common.BytesToAddress([]byte{0x01, 0x11})},
-		},
-		L1GasUsed:  gasUsed,
-		L1GasPrice: gasUsed,
-		L1Fee:      gasUsed,
-		FeeScalar:  "6",
-	}
-
-	data, err := rlp.EncodeToBytes(receipt)
-	require.NoError(t, err)
-	var result storedReceiptRLP
-	err = rlp.DecodeBytes(data, &result)
-	require.NoError(t, err)
-	require.Equal(t, receipt.L1GasUsed, result.L1GasUsed)
-	require.Equal(t, receipt.L1GasPrice, result.L1GasPrice)
-	require.Equal(t, receipt.L1Fee, result.L1Fee)
-	require.Equal(t, receipt.FeeScalar, result.FeeScalar)
 }
 
 func TestDeriveLogFields(t *testing.T) {
